+++ src/main.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './app/App';
 import './index.css';
-import ErrorBoundary from './ErrorBoundary';
+import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary';
 import { BrowserRouter } from 'react-router-dom';
-import { PokemonProvider } from './PokemonContext';
+import { PokemonProvider } from './context/PokemonContext';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
